requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
psycopg2-binary  
pydantic[email]
uuid
datetime
typing
passlib[bcrypt]==1.7.4
PyJWT
python-multipart
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv